+++ packages/elements/avatar/src/hoc/mapProps.js
@@ -7,7 +7,7 @@ export default function mapProps(mapping: {}) {
   return (DecoratedComponent: ComponentType) =>
     // TODO: type this correctly
     class MapProps extends Component<*> {
-      static displayName: string = getDisplayName(
+      static displayName: string | void | null = getDisplayName(
         'mapProps',
         DecoratedComponent,
       );