+++ packages/fabric/media-avatar-picker/src/avatar-picker-dialog/styled.tsx
@@ -7,8 +7,14 @@ export const AvatarPickerViewWrapper = styled.div`
 
 export const ModalHeader = styled.div`
   margin: 15px;
+  font-weight: 500;
 `;
 
 export const CroppingWrapper = styled.div`
   display: inline-block;
+`;
+
+export const ModalFooterButtons = styled.div`
+  text-align: right;
+  width: 100%;
 `;