+++ packages/fabric/editor-bitbucket-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-bitbucket-transformer
 
+## 2.1.8
+
 ## 2.1.7
 
 ## 2.1.6